Refactor the tracked lock code and extend the API.
Refactor the lock tracking code so that locks are not acquired or released
by the code that does the tracking.  Instead, the tracking code is called
after the locks is acquired or before the lock is released.

Expose the interfaces to the lock tracking code so that DLM locks, and
potentially the chunked put connection code can call them for tracking.

Now the wrappers for sync.Mutex and sync.RWMutex are in api.go, saving
an extra argument to the tracker routines, an interface class and a set
of wrapper routines for dlm.RWLockStruct that were too clever by half.

This may be a slight performance hit for contended locks, but probably
not much.

Write a bunch of unit tests.

Lock tracking is disabled by default..